An enrichResult object for differentially expressed genes in the cell line comparison.

An enrichResult object for differentially expressed genes in the cell line comparison.

Format

An enrichResult object, generated with the enrichGO function from the clusterProfiler package.

Details

This enrichResult object was created to test for functional enrichment using the GO Biological Process (BP) ontology on the top 100 differentially expressed genes from the cell line comparison.

Details on how this object has been created are included in the create_carnation_data.R script, included in the scripts folder of the Carnation package.

References

Himes BE, Jiang X, Wagner P, Hu R, Wang Q, Klanderman B, Whitaker RM, Duan Q, Lasky-Su J, Nikolos C, Jester W, Johnson M, Panettieri R Jr, Tantisira KG, Weiss ST, Lu Q. “RNA-Seq Transcriptome Profiling Identifies CRISPLD2 as a Glucocorticoid Responsive Gene that Modulates Cytokine Function in Airway Smooth Muscle Cells.” PLoS One. 2014 Jun 13;9(6):e99625. PMID: 24926665. GEO: GSE52778